refactor(api): move ai-agents to modules/agents/core and share the domain params
