+++ apps/api/src/app.ts
@@ -12,7 +12,7 @@ import { planner } from './planner';
 import { webhookTestRoutes } from './webhook-test/routes';
 import { mountMcp } from './mcp/mount';
 import { setMcpApp } from './mcp/app-ref';
-import { internalAgentRunRoutes } from './ai-agents/internal-routes';
+import { internalAgentRunRoutes } from './modules/agents/core/internal-routes';
 import { internalNotificationRoutes } from './notifications/internal-routes';
 import { internalTelegramRoutes } from './telegram/internal-routes';
 import { githubWebhookRoutes } from './github/routes';
