+++ apps/api/src/modules/agents/core/__tests__/integration/agent-runs.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from 'bun:test';
-import { authedApi, type Api } from '../../../__tests__/helpers/app';
-import { signUpTestUser } from '../../../__tests__/helpers/auth';
-import { resetDb } from '../../../__tests__/helpers/db';
+import { authedApi, type Api } from '#tests/helpers/app';
+import { signUpTestUser } from '#tests/helpers/auth';
+import { resetDb } from '#tests/helpers/db';
 
 // The run-history endpoint: GET /projects/:key/ai-agents/:agentId/runs lists an agent's
 // triggered runs (a mention or a delegation), newest first, keyset-paginated. Runs are
