+++ apps/api/src/modules/agents/core/__tests__/integration/ai-agents.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach } from 'bun:test';
-import { authedApi, type Api } from '../../../__tests__/helpers/app';
-import { signUpTestUser } from '../../../__tests__/helpers/auth';
-import { resetDb } from '../../../__tests__/helpers/db';
-import { untaggedRoutes } from '../../../__tests__/helpers/mcp';
+import { authedApi, type Api } from '#tests/helpers/app';
+import { signUpTestUser } from '#tests/helpers/auth';
+import { resetDb } from '#tests/helpers/db';
+import { untaggedRoutes } from '#tests/helpers/mcp';
 
 // AI agents attached to a project. Each agent is backed by a hidden bot user, owns a
 // better-auth API key, and is a project member acting under a project role. An
