+++ apps/api/src/modules/agents/core/__tests__/integration/thread-cleanup.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach } from 'bun:test';
 import { db } from '@repo/db';
 import { sql } from 'drizzle-orm';
-import { authedApi, type Api } from '../../../__tests__/helpers/app';
-import { signUpTestUser } from '../../../__tests__/helpers/auth';
-import { resetDb } from '../../../__tests__/helpers/db';
+import { authedApi, type Api } from '#tests/helpers/app';
+import { signUpTestUser } from '#tests/helpers/auth';
+import { resetDb } from '#tests/helpers/db';
 import { ensureThread, buildMemory } from '../../runtime/memory';
 
 // Agent conversation threads live in Mastra's tables, which carry no foreign key of
