+++ apps/api/src/modules/agents/core/__tests__/unit/mentions.test.ts
