+++ apps/api/src/modules/agents/core/__tests__/unit/tools.test.ts
