+++ apps/api/src/modules/agents/core/helpers/errors.ts
