+++ apps/api/src/modules/agents/core/issue-query.ts
