+++ apps/api/src/modules/agents/core/mentions.ts
