+++ apps/api/src/modules/agents/core/runtime/tools/catalog.ts
